Address some issues with NuGet package creation
* Remove netcore as a target framework from build.ps1

The CloudinaryDotNet.Core.csproj doesn't designate a .Net Core TFM (target framework moniker) so the actual build of this project only produces .NetStandard 1.3.

* Explicitly define target framework dependencies.

This means that all the Nuget tools will report the correct dependencies.

Additionally, only include files in the nuget package that have a file name that starts with "Cloudinary" except if the file name ends with ".deps.json". The side effect of this means that when developers consume of this nuget package, especially those targeting .Net Framework, it doesn't highjack the version/path of Newtonsoft.Json that their projects are already referencing.